requirements.txt
flask
pymongo
mongoengine
werkzeug
python-dotenv